rework callback generation
