+++ demo_packet.go
@@ -80,7 +80,7 @@ func (p *Parser) onCDemoPacket(m *dota.CDemoPacket) error {
 
 	// Dispatch messages in order, returning on handler error.
 	for _, m := range ms {
-		if err := p.CallByPacketType(m.t, m.buf); err != nil {
+		if err := p.Callbacks.callByPacketType(m.t, m.buf); err != nil {
 			return err
 		}
 	}
